refactor: 서비스 로직 도메인에 위임
