+++ src/main/java/clofi/codeython/judge/domain/creator/RuntimeEnvironmentCreator.java
@@ -1,0 +1,7 @@
+package clofi.codeython.judge.domain.creator;
+
+import java.util.List;
+
+public interface RuntimeEnvironmentCreator {
+    void config(List<String> inputTypes, String code, String route);
+}
